+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * [ObjectServer] Added `ConnectionState` enum describing the states a connection can be in.
 * [ObjectServer] Added `SyncSession.isConnected()`.
 * [ObjectServer] Added support for observing connection changes for a session using `SyncSession.addConnectionChangeListener()` and `SyncSession.removeConnectionChangeListener()`.
+* Added support for `RealmQuery.limit()` (#544).
 
 ### Bug Fixes